Add files via upload
Project2 - education center. The program asks for a file that contains the course center for the course information: theme, course, location and number of participants. 
Data storage happens to the selected data structure, after which the user can enter search commands for it. If the file cannot be opened or the file is not found four fields, an error message is printed and the program execution ends. 
There are eight different search formats: 'themes', 'courses <location> <theme>, 'available', 'courses_in_theme <theme>', 'favorite theme', 'courses_in_location_ <location>', 'cancel <course>' and 'quit'. 
The program execution is terminated when the user enters the command quit. Of the wrong ones an error message is printed from the inputs, but the program continues after this. 
* Command 'themes' prints themes alphabetically. 
* Command 'courses <place>' prints in alphabetical order according to the name all courses and number of participants that can be found in that theme and from the locality. 
* Command 'available' prints the courses that are not available not full, i.e. the number of participants is less than 50. 
* Command 'courses_in_theme <theme>' prints courses according to the theme from all locations in it rules when courses are given in file. 
* Command 'favorite theme' prints the theme of the recommendation with the largest number of participants. 
* Command 'cancel <course>' command to cancel course from all locations. Error checking is done for inputs <command>, <locality>, <theme>. If granted in incorrect format or not found in file, an error message is printed.